refactor: 포인트에 대한 처리를 hook으로 분리
